phish-api/requirements.txt
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cmake==3.22.5
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
dlib==19.24.0
entrypoints==0.4
executing==1.0.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.37.1
gunicorn==20.1.0
idna==3.4
imutils==0.5.4
ipykernel==6.15.2
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.5
nltk==3.7
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.4
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pygame==2.1.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==23.2.1
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
scipy==1.8.1
seaborn==0.12.0
six==1.16.0
sklearn==0.0
stack-data==0.5.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
xgboost==1.7.1
